web/client: enhanced lib/app code split
