+++ web/client/src/App.styles.tsx
@@ -1,9 +1,10 @@
-import { Theme, Breakpoint, Paper, Container } from '@mui/material';
+import { Theme, Paper, Container } from '@mui/material';
 import { styled } from '@mui/styles';
-import { CreateCSSProperties } from '@mui/styles/withStyles';
 
 export const StyledAppFlexDiv = styled('div')(() => {
-  return { display: 'flex' };
+  return {
+    maxWidth: '1920px',
+  };
 });
 
 export const StyledAppApiLoading = styled('div')({
@@ -13,21 +14,24 @@ export const StyledAppApiLoading = styled('div')({
 export const StyledAppContent = styled('main')({
   flexGrow: 1,
   flexFlow: 'row wrap',
-  height: '100vh',
-  overflow: 'auto'
+  overflow: 'auto',
 });
 
-export const StyledAppBarSpacer = styled('div')(({ theme }: { theme: Theme }) => {
-  return (theme.mixins.toolbar as CreateCSSProperties);
-});
-
-export const StyledAppPaper = styled(Paper)(({ theme }: { theme: Theme }) => {
-  return {
-    padding: theme.spacing(2),
-    overflow: 'auto',
-    maxWidth: 'none'
-  };
-});
+export const StyledAppPaper = styled(
+  (props:any): JSX.Element => {
+    return (
+      <Paper elevation={1}>{props.children}</Paper >
+    );
+  }
+)(
+  ({ theme }: { theme: Theme }) => {
+    return {
+      padding: theme.spacing(2),
+      overflow: 'auto',
+      maxWidth: 'none'
+    };
+  }
+);
 
 interface ContainerWrapperProps {
   children: JSX.Element | JSX.Element[],
@@ -35,7 +39,7 @@ interface ContainerWrapperProps {
 }
 const ContainerWrapper = (props: ContainerWrapperProps): JSX.Element => {
   return (
-    <Container className={props.className}>{props.children}</Container >
+    <Container maxWidth={false} className={props.className}>{props.children}</Container >
   );
 }
 
@@ -43,7 +47,6 @@ export const StyledContainer = styled(ContainerWrapper)(({ theme }: { theme: The
   return {
     padding: theme.spacing(2),
     marginLeft: 0,
-    maxWidth: '1920px',
   }
 });
