+++ web/client/src/lib/components/index.ts
@@ -2,5 +2,5 @@ export { default as Create } from './Create';
 export { default as Edit } from './Edit';
 export { default as List } from 'lib/components/List/List';
 export { default as View } from './View';
-export { default as Login } from './Login';
+export { default as Login } from '../../components/Login';
 export * from 'lib/components/layout';
